# docs/swagger.yaml
definitions:
...
  v1.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
  v1.UpdateUserGeoRequest:
    properties:
      address:
        type: string
      first_area_id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      second_area_id:
        type: integer
      third_area_id:
        type: integer
    type: object
  v1.UpdateUserInfoRequest:
    properties:
      avatar:
        type: string
      name:
        type: string
      phone:
        type: string
      sex:
        type: integer
    type: object
...
  v1.UserInfoResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/v1.UserInfoResponseData'
      message:
        type: string
      trace_id:
        type: string
    type: object
  v1.UserInfoResponseData:
    properties:
      avatar:
        type: string
      contact_voucher_num:
        type: integer
      first_top_status:
        type: integer
      name:
        type: string
      new_customer_status:
        type: integer
      phone:
        type: string
      profile_complete_status:
        type: integer
      sex:
        type: integer
      share_refresh_available:
        type: integer
      share_refresh_next_available_at:
        type: string
      share_refresh_remaining_seconds:
        type: integer
      user_code:
        type: string
      user_id:
        type: integer
    type: object
  v1.UserInviteItem:
    properties:
      avatar:
        type: string
      consume_status:
        type: integer
      create_at:
        type: string
      login_status:
        type: integer
      name:
        type: string
      publish_status:
        type: integer
      user_id:
        type: integer
      voucher_earned:
        type: integer
    type: object
  v1.UserInviteListRequest:
    properties:
      page_num:
        type: integer
      page_size:
        type: integer
    type: object
  v1.UserInviteListResponseData:
    properties:
      consume_voucher_reward:
        type: integer
      invite_total:
        type: integer
      list:
        items:
          $ref: '#/definitions/v1.UserInviteItem'
        type: array
      login_voucher_reward:
        type: integer
      publish_voucher_reward:
        type: integer
      total:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Nunu Example API
  version: 1.0.0
paths:
...
  /user/info:
    get:
      consumes:
      - application/json
      description: 返回当前登录用户的基本信息。share_refresh_available=1表示可分享免费刷新；冷却中同时返回下次可用时间和剩余秒数。冷却时长为24小时。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserInfoResponse'
      security:
      - Bearer: []
      summary: 查询个人信息
      tags:
      - 用户模块
  /user/invites:
    post:
      consumes:
      - application/json
      description: 返回当前用户邀请注册的用户列表，按注册时间倒序。invite_total 为累计邀请总人数；total 为当前分页总数。
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UserInviteListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserInviteListResponseData'
      security:
      - Bearer: []
      summary: 邀请记录
      tags:
      - 个人中心
...
  /user/update/geo:
    post:
      consumes:
      - application/json
      description: 更新用户的地理位置。所有字段均为可选指针，传哪个改哪个。通常在小程序授权地理位置后调用。
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateUserGeoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      security:
      - Bearer: []
      summary: 更新位置信息
      tags:
      - 用户模块
  /user/update/info:
    post:
      consumes:
      - application/json
      description: '更新用户基本信息。所有字段均为可选指针，传哪个改哪个。sex: 1=男 2=女。phone 更新后 profile_complete_status
        自动置1。'
      parameters:
      - description: params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateUserInfoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.Response'
      security:
      - Bearer: []
      summary: 更新个人信息
      tags:
      - 用户模块
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"